fix text implementation
git-svn-id: https://osmand.googlecode.com/svn/trunk@516 e29c36b1-1cfa-d876-8d93-3434fc2bb7b8

+++ DataExtractionOSM/src/net/osmand/osm/MapRenderObject.java
@@ -48,7 +48,7 @@ public class MapRenderObject {
 		return Algoritms.parseIntFromBytes(data, ind * 8);
 	}
 	
-	public float getPoint31XTile(int ind) {
+	public int getPoint31XTile(int ind) {
 		return Algoritms.parseIntFromBytes(data, ind * 8 + 4);
 	}
